+++ src/redux/api.ts
@@ -4,6 +4,32 @@ import Cookies from 'js-cookie';
 import type { AxiosResponse } from 'axios';
 
 const baseURL = process.env.NEXT_PUBLIC_API_URL;
+if (baseURL === undefined) {
+  throw new Error('NEXT_PUBLIC_API_URL is undefined');
+}
+
+export const get = async <A, RT>(
+  path: string,
+  params: A,
+  withToken = true,
+): Promise<AxiosResponse<RT, any>> => {
+  let token: string | undefined;
+  if (withToken) {
+    token = Cookies.get('token');
+  }
+  const url = `${baseURL}/${path}`;
+  return await axios<RT>({
+    method: 'GET',
+    url,
+    params,
+    headers:
+      token !== undefined
+        ? {
+            Authorization: `Bearer ${token}`,
+          }
+        : undefined,
+  });
+};
 
 export const post = async <A, RT>(
   path: string,
@@ -19,9 +45,12 @@ export const post = async <A, RT>(
     method: 'POST',
     url,
     data: body,
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
+    headers:
+      token !== undefined
+        ? {
+            Authorization: `Bearer ${token}`,
+          }
+        : undefined,
   });
 };