+++ src/redux/generalActions.ts
@@ -3,7 +3,7 @@ import { AxiosError } from 'axios';
 
 import { getRefreshToken, getToken, isTokenExpired } from '../utils/auth';
 
-import { post, requestNewToken } from './api';
+import { get, post, requestNewToken } from './api';
 
 import type { KnownError } from './knownError';
 import type { AsyncThunk } from '@reduxjs/toolkit';
@@ -42,6 +42,8 @@ const generateRequest = <RT = unknown, A = void>(
         switch (method) {
           case 'post':
             return (await post<A, RT>(path, params, options.withToken)).data;
+          case 'get':
+            return (await get<A, RT>(path, params, options.withToken)).data;
           default:
             return null as RT;
         }