+++ src/redux/store.ts
@@ -1,9 +1,11 @@
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 
+import alarmsReducer from './slices/alarm/alarmsSlice';
 import authReducer from './slices/auth/authSlice';
 
 const rootReducer = combineReducers({
   auth: authReducer,
+  alarm: alarmsReducer,
 });
 
 const store = configureStore({