+++ src/utils/spanishDays.ts
@@ -1,0 +1,11 @@
+import { type WeekDay } from '../redux/slices/alarm/alarmsSlice';
+
+export const spanishDays: Record<WeekDay, string> = {
+  monday: 'Lunes',
+  tuesday: 'Martes',
+  wednesday: 'Miércoles',
+  thursday: 'Jueves',
+  friday: 'Viernes',
+  saturday: 'Sábado',
+  sunday: 'Domingo',
+};