+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Types of changes
+
+- **Added** for new features.
+- **Changed** for changes in existing functionality.
+- **Deprecated** for soon-to-be removed features.
+- **Removed** for now removed features.
+- **Fixed** for any bug fixes.
+- **Security** in case of vulnerabilities.
+
+## [0.2.0] - 2024-01-13
+
+### Added
+
+- useCheckLoging hook to check if user is logged in
+- General Layout
+- Alarm state
+- Types of changes in changelog
+- Time picker from MUI X
+- Set Alarms in Frontend
+
+### Removed
+
+- Pene actions and slice
+
 ## [0.1.0] - 2024-01-13
 
 ### Added